+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-Basic Fluid Hopper v0.5.1-1.21.1
+## Milk Edition!
 
-* Actually saving/loading configuration!
+- Milk (added on Fabric / enabled on NeoForge)
+    - Milk is technically placeable, but milk bucket does not place milk
+    - To actually have flowing milk, you'll need to use the /replace command or install another mod that can place fluids
+- Cows standing on a fluid hopper will be automatically milked
+- Bottles can be filled with milk directly from a hopper or with a dispenser
+    - Milk bottles are consumable just like milk buckets
+    - You still need a bucket to milk a cow. Trying to milk a cow into a bottle sounds difficult and unpleasant.
+
+## Other changes
+* Fixed invisible water in cauldron (Fabric)
+* Cauldron can be filled by rain (water) or dripstone (water or lava)